IO-143: Fix extension uninstallation/installation
