--- CRM/CiviAwards/Setup/DeleteApplicantReviewCustomField.php
@@ -1,25 +1,0 @@
-<?php
-
-/**
- * Class DeleteApplicantReviewCustomField.
- */
-class CRM_CiviAwards_Setup_DeleteApplicantReviewCustomField {
-
-  /**
-   * Deletes the Applicant review custom group set.
-   */
-  public function apply() {
-    $result = civicrm_api3('CustomGroup', 'get', [
-      'name' => 'Applicant_Review',
-    ]);
-
-    if (empty($result['id'])) {
-      return;
-    }
-
-    civicrm_api3('CustomGroup', 'delete', [
-      'id' => $result['id'],
-    ]);
-  }
-
-}
